Retrofit the API structure
