+++ src/main/java/com/billyyccc/api/utils/RestApiUtil.java
@@ -1,4 +1,4 @@
-package com.billyyccc.http.utils;
+package com.billyyccc.api.utils;
 
 import io.vertx.core.json.DecodeException;
 import io.vertx.core.json.Json;
