+++ src/test/java/com/billyyccc/api/book/DeleteBookByIdHandlerTest.java
@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2017 Billy Yuan
+ * Copyright (c) 2018 Billy Yuan
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -22,19 +22,19 @@
  * SOFTWARE.
  */
 
-package com.billyyccc.http.handler;
+package com.billyyccc.api.book;
 
 import com.billyyccc.database.reactivex.BookDatabaseService;
-import com.billyyccc.entity.Book;
-import io.reactivex.Single;
-import io.vertx.core.json.JsonArray;
-import io.vertx.core.json.JsonObject;
+import com.billyyccc.api.handler.BookApis;
+import io.reactivex.Completable;
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
 import io.vertx.ext.unit.junit.RunTestOnContext;
 import io.vertx.ext.unit.junit.VertxUnitRunner;
 import io.vertx.reactivex.core.Vertx;
+import io.vertx.reactivex.core.http.HttpClient;
 import io.vertx.reactivex.ext.web.Router;
+import io.vertx.reactivex.ext.web.handler.BodyHandler;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Rule;
@@ -43,13 +43,13 @@ import org.junit.runner.RunWith;
 import org.mockito.Mockito;
 
 /**
- * This test Class is to perform unit tests for GetBooksHandler of books.
+ * This test Class is to perform unit tests for DeleteBookByIdHandler of books.
  *
  * @author Billy Yuan <billy112487983@gmail.com>
  */
 
 @RunWith(VertxUnitRunner.class)
-public class GetBooksHandlerTest {
+public class DeleteBookByIdHandlerTest {
   @Rule
   public RunTestOnContext rule = new RunTestOnContext();
   private Vertx vertx;
@@ -61,24 +61,15 @@ public class GetBooksHandlerTest {
 
     BookDatabaseService bookDatabaseService = Mockito.mock(BookDatabaseService.class);
 
-    JsonArray mockDbResponse = new JsonArray()
-      .add(new JsonObject()
-        .put("id", 1)
-        .put("title", "Effective Java")
-        .put("category", "java")
-        .put("publicationDate", "2009-01-01"))
-      .add(new JsonObject()
-        .put("id", 2)
-        .put("title", "Thinking in Java")
-        .put("category", "java")
-        .put("publicationDate", "2006-02-20"));
+    int bookId = 1;
 
+    Mockito.when(bookDatabaseService.rxDeleteBookById(bookId)).thenReturn(Completable.complete());
 
-    Mockito.when(bookDatabaseService.rxGetBooks(new Book())).thenReturn(Single.just(mockDbResponse));
-
-    router.get("/books").handler(new GetBooksHandler(bookDatabaseService));
+    router.route().handler(BodyHandler.create());
+    router.delete("/books/:id").handler(BookApis.deleteBookByIdHandler(bookDatabaseService));
 
     vertx.createHttpServer().requestHandler(router::accept).listen(1234, testContext.asyncAssertSuccess());
+
   }
 
   @After
@@ -88,28 +79,22 @@ public class GetBooksHandlerTest {
 
   @Test
   public void restApiTest(TestContext testContext) {
+    HttpClient httpClient = vertx.createHttpClient();
+
     Async async = testContext.async();
-    vertx.createHttpClient().getNow(1234, "localhost", "/books", res -> {
+
+    httpClient.delete(1234, "localhost", "/books/1", res -> {
+
       testContext.assertEquals(200, res.statusCode());
+
       res.bodyHandler(body -> {
-        testContext.assertTrue(body.length() > 0);
+        testContext.assertTrue(body.length() == 0);
 
-        JsonArray expectedResponseBody = new JsonArray()
-          .add(new JsonObject()
-            .put("id", 1)
-            .put("title", "Effective Java")
-            .put("category", "java")
-            .put("publicationDate", "2009-01-01"))
-          .add(new JsonObject()
-            .put("id", 2)
-            .put("title", "Thinking in Java")
-            .put("category", "java")
-            .put("publicationDate", "2006-02-20"));
-
-        testContext.assertEquals(expectedResponseBody, body.toJsonArray());
+        httpClient.close();
         async.complete();
       });
-    });
+    }).putHeader("Content-Type", "application/json; charset=utf-8")
+      .end();
   }
 
 }
